backend/requirements.txt
google-adk
openai
pinecone
python-dotenv

langchain
langchain-openai
langchain-pinecone
uvicorn
fastapi
